moved Fragment to support fragment in themee browser
